Add files via upload
Here is the updated database create script along with the updated ER diagram. I also included a couple scripts that calculate the student's class average, convert the average to a 4.00 gpa scale, and then calculates the semester average and stores it into the studentGpa table. I will make these either stored procedures or triggers when we implement the GUI.