Add files via upload
Visual Studio creation of rough prototype for GUI. This is done in Visual C#, but Ben and Brian are going to build one using Visual Basic. Feel free to use this as a guideline. How it will work:  User logs in via a login form. Based on the credentials, either a student form or an admin form will generate. The student form is a read only form with the student info and GPA on the main form. The student form has a panel on the right. From the registered courses generated on the left, when a student clicks on one of those courses, his information about that course appears dynamically in the panel on the right...the course name, average, and gpa. The exams and the grades for those exams are displayed in the listbox below.

I created an MDI Form for the admin and put a dataGridView in it to populate data based on the menu option selected. There should be an ability to search based on the text box fields entered, and there is a second button to update changes made to the dataViewGrid. I spent