Add files via upload
I may have slightly altered the database so I included another backup file. The SIMS(4-28).zip contains the solution to the C# GUI I'm working on. I have two panels for the admin side up and running. I just need to create three more. It will be a lot of copy and pasting and creating a few more sql procedures.